+++ Core/Texture.h
@@ -4,7 +4,7 @@ class Texture
 {
 public:
     Texture(ID3D12Device* device, DescriptorHeap* descHeap);
-    void Upload(ResourceUploadBatch* resUpload, uint32_t descHeapIdx, const wstring& filename);
+    void Upload(ResourceUploadBatch* resUpload, size_t descHeapIdx, const wstring& filename);
 
     XMUINT2 GetSize() const noexcept;
     void Reset();
@@ -14,6 +14,6 @@ private:
     ID3D12Device* m_device;
     DescriptorHeap* m_descHeap;
     Microsoft::WRL::ComPtr<ID3D12Resource> m_texture;
-    uint32_t m_descHeapIdx{ 0 };
+    size_t m_descHeapIdx{ 0 };
 };
